+++ test/libsolidity/syntaxTests/constants/cyclic_dependency_3.sol
@@ -5,7 +5,7 @@ contract C {
     uint constant c = b;
 }
 // ----
-// TypeError: The value of the constant x has a cyclic dependency via a.
-// TypeError: The value of the constant a has a cyclic dependency via b.
-// TypeError: The value of the constant b has a cyclic dependency via c.
-// TypeError: The value of the constant c has a cyclic dependency via b.
+// TypeError: (17-36): The value of the constant x has a cyclic dependency via a.
+// TypeError: (42-65): The value of the constant a has a cyclic dependency via b.
+// TypeError: (71-90): The value of the constant b has a cyclic dependency via c.
+// TypeError: (96-115): The value of the constant c has a cyclic dependency via b.