+++ test/libsolidity/syntaxTests/inheritance/duplicated_constructor_call/ancestor_V050.sol
@@ -4,4 +4,4 @@ contract A { constructor(uint) public { } }
 contract B is A(2) { constructor() public {  } }
 contract C is B { constructor() A(3) public {  } }
 // ----
-// DeclarationError: Base constructor arguments given twice.
+// DeclarationError: (156-160): Base constructor arguments given twice.