+++ test/libsolidity/syntaxTests/inheritance/duplicated_constructor_call/base_multi_no_constructor.sol
@@ -3,4 +3,4 @@ contract A is C(2) {}
 contract B is C(2) {}
 contract D is A, B {}
 // ----
-// Warning: Base constructor arguments given twice.
+// Warning: (87-108): Base constructor arguments given twice.