+++ test/libsolidity/syntaxTests/inheritance/duplicated_constructor_call/base_multi_no_constructor_modifier_style.sol
@@ -3,4 +3,4 @@ contract A is C { constructor() C(2) public {} }
 contract B is C { constructor() C(2) public {} }
 contract D is A, B { }
 // ----
-// Warning: Base constructor arguments given twice.
+// Warning: (141-163): Base constructor arguments given twice.